Search Functionality Complete
The search function works and, while not entirely optimized, does correctly and accurately return the records that adhere to the paramters the user specified.

BUT, not everything has been error checked yet. For example, if the user specifies a minimum amount that is greater than the specified maximum amount, this is not caught. Same goes for the specified date range.

The EditRecordWindow works as well, and does in fact mark records for deletion, but when taken back to the results the visual is not updated. TODO