Record Modification and Database Updating (Working Midterm Version)
All records the user modifies, are stored and redisplayed if they execute multiple searches while at the same window. When travelling back to the user dashboard, the modified records are taken into account when displaying month record ratios.

Midterm down, got a lot of work to do for the final.