Dashboard Changes / Early Tutorial Stages for Dashboard
While my program works, I had someone try and break it and they broke it, wide open. There are quite a few validation problems, like negative signs being accepted and commas being accepted too. The biggest validation problem is apostrophes being accepted in the record description, as this causes an SQL expection when inserting or updating the record in the database.

These validation problems are not the end of the world, its just a pain because there are quite a few.

I have also started to implement a simple Help button on the dashboard that essentially executes a simple tutorial procedure. This works on the dashboard, but I need to implement the same thing on almost all other interfaces. (Now that I know how to do it, it shouldnt be super difficult.)