Semi-Final Validation Checks in place/Tutorial Sequences in place/Finishing Touches for demo
There are much better validation checks in place at the most important input sites; mainly any amount inputs, and descriptions.

Amount inputs are stripped of any '-' characters present before testing to see if it can be converted into a double. This check also prevents the user from entering a $ or something (really, anything that isnt a number OR ',' OR '.'). The check does accept amounts input in the format      ##,##     ##.##      ##,###.##      and all variations. Just as long as the input can be parsed into a double, then it is valid.

The description inputs are now automatically stripped of any and all charaters that are not alpha numberic or whitespace. Hooray for no more SQL expections (I really hope so)