Small Changes to Default color settings
Inserted small check to see if this is the first time the user is using the program. On registration, their color preferences haven't been saved. So when they log in for the first time, unless a certain element of the UI has its color hardcoded in the XAML, then the color is white. While this might be a slight oversight in the registration process, all this does is make sure that some default value is loaded. In the database table, the color and currency preferences can be null. When the user logs out, their color preferences are saved to the database as well. Perhaps a small boolean check to see if they have changed their preferences can be put in place, just so that a users color preferences and what not are not updated EVERY SINGLE TIME that they log out. This is small scale, but could be a small optimization, database wise.