Laying Foundation for new Record objects
So far, have created the classes for Expense and Income records. Instead of dealing with NewIncomeRecord, ExistingIncomeRecord, and SearchResultRecord for dealing with the different stages of record keeping in the program, I am condensing all three possible classes into a single, easy to manage class.

ResultsUtility will now contain almost all of the functions that find and sort records based on the search parameters that the user has specified. For now, I am just trying to copy everything over, more or less, to new files, and then begin to integrate the new classes into the program. Hence the new branch, because things could get messy. While this is a lot of work, this will make the entire project much easier on the eyes and it will make more sense.

tl;dr creating new classes for records and essentially rewriting the searching functions