More functions for ResultsUtility.cs
Created implementations for the GetExpenseSearchResults and GetIncomeSearchResults functions from SearchResults.cs

These implemetations are drastically simplified in comparison because of the revisions with IncomeRecord and ExpenseRecord, instead of multiple types of records for each.

These have not been tested so far though, so I do not know how well, if at all, they work. But they should... We will see.