Transformative Process Started
Now that the ExpenseRecord and IncomeRecord models have been defined and hashed out, the process of refactoring important code segments has begun.

When it comes to SearchRecordsWindow.xaml.cs, I think for the most part the important segments have been refactored to utilize the new design.

ResultsUtility.cs had some silly errors in it, so those were fixed.

User.cs now has ExpenseRecord and IncomeRecord list properties.

Database.cs is the next biggest file to refactor. At the moment, none of the records from the database are being read into new object types, and the function to update the database with new or modified records does not look at the new obejct types. These are the biggest steps to be changed in order to actually test the fundamental changes made so far.